requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.10.1+cu113
torchvision==0.11.2+cu113

numpy==1.21.6
scipy==1.7.3
Pillow==7.1.2
scikit-image==0.19.2
networkx==2.3
matplotlib==3.1.0

pytest==7.1.3
mypy==0.982
pytest-mypy==0.10.0
coverage==6.5.0
